+++ src/components/Filter/FilterContent/FilterContentBody.tsx
@@ -1,12 +1,10 @@
-import { FormControlLabel, Radio, TextField } from "@material-ui/core";
-import { fade } from "@material-ui/core/styles/colorManipulator";
+import { FormControlLabel, Radio } from "@material-ui/core";
 import FormSpacer from "@saleor/components/FormSpacer";
 import { MultiAutocompleteChoiceType } from "@saleor/components/MultiAutocompleteSelectField";
 import SingleSelectField from "@saleor/components/SingleSelectField";
-import { makeStyles } from "@saleor/theme";
 import classNames from "classnames";
 import React from "react";
-import { FormattedMessage, useIntl } from "react-intl";
+import { useIntl } from "react-intl";
 
 import Arrow from "../Arrow";
 import FilterAutocompleteField, {
@@ -15,45 +13,14 @@ import FilterAutocompleteField, {
 import FilterOptionField from "../FilterOptionField";
 import { FilterReducerAction } from "../reducer";
 import { FieldType, FilterType, IFilterElement } from "../types";
-import { getIsFilterMultipleChoices } from "./utils";
-
-const useStyles = makeStyles(
-  theme => ({
-    andLabel: {
-      margin: theme.spacing(0, 2)
-    },
-    arrow: {
-      marginRight: theme.spacing(2)
-    },
-    filterSettings: {
-      background: fade(theme.palette.primary.main, 0.1),
-      padding: theme.spacing(2, 3)
-    },
-    input: {
-      padding: "12px 0 9px 12px"
-    },
-    inputRange: {
-      alignItems: "center",
-      display: "flex"
-    },
-
-    option: {
-      left: -theme.spacing(0.5),
-      position: "relative"
-    },
-    optionRadio: {
-      left: -theme.spacing(0.25)
-    }
-  }),
-  { name: "FilterContentBody" }
-);
-
-const filterTestingContext = "filter-field";
+import FilterRangeField from "./FilterRangeField";
+import FilterTextField from "./FilterTextField";
+import useStyles from "./styles";
+import { filterTestingContext, getIsFilterMultipleChoices } from "./utils";
 
 export interface FilterContentBodyProps<T extends string = string> {
-  children?: React.ReactNode;
   filter: IFilterElement<T>;
-  currencySymbol?: string;
+  currencySymbol: string;
   initialAutocompleteDisplayValues: FilterAutocompleteDisplayValues;
   onFilterPropertyChange: React.Dispatch<FilterReducerAction<T>>;
   autocompleteDisplayValues: FilterAutocompleteDisplayValues;
@@ -78,28 +45,10 @@ const FilterContentBody: React.FC<FilterContentBodyProps> = ({
     <div className={classes.filterSettings}>
       {children}
       {filter.type === FieldType.text && (
-        <TextField
-          data-test={filterTestingContext}
-          data-test-id={filter.name}
-          fullWidth
-          name={filter.name}
-          InputProps={{
-            classes: {
-              input: classes.input
-            }
-          }}
-          value={filter.value[0]}
-          onChange={event =>
-            onFilterPropertyChange({
-              payload: {
-                name: filter.name,
-                update: {
-                  value: [event.target.value, filter.value[1]]
-                }
-              },
-              type: "set-property"
-            })
-          }
+        <FilterTextField
+          currencySymbol={currencySymbol}
+          filter={filter}
+          onFilterPropertyChange={onFilterPropertyChange}
         />
       )}
       {[FieldType.date, FieldType.price, FieldType.number].includes(
@@ -112,7 +61,7 @@ const FilterContentBody: React.FC<FilterContentBodyProps> = ({
             value={filter.multiple ? FilterType.MULTIPLE : FilterType.SINGULAR}
             InputProps={{
               classes: {
-                input: classes.input
+                input: classes.fieldInput
               }
             }}
             onChange={event =>
@@ -133,101 +82,16 @@ const FilterContentBody: React.FC<FilterContentBodyProps> = ({
               <Arrow className={classes.arrow} />
             </div>
             {filter.multiple ? (
-              <>
-                <TextField
-                  data-test={filterTestingContext}
-                  data-test-id={filter.name}
-                  data-test-range-type="min"
-                  fullWidth
-                  name={filter.name + "_min"}
-                  InputProps={{
-                    classes: {
-                      input: classes.input
-                    },
-                    endAdornment:
-                      filter.type === FieldType.price && currencySymbol,
-                    type: filter.type === FieldType.date ? "date" : "number"
-                  }}
-                  value={filter.value[0]}
-                  onChange={event =>
-                    onFilterPropertyChange({
-                      payload: {
-                        name: filter.name,
-                        update: {
-                          value: [event.target.value, filter.value[1]]
-                        }
-                      },
-                      type: "set-property"
-                    })
-                  }
-                />
-                <span className={classes.andLabel}>
-                  <FormattedMessage
-                    defaultMessage="and"
-                    description="filter range separator"
-                  />
-                </span>
-                <TextField
-                  data-test={filterTestingContext}
-                  data-test-id={filter.name}
-                  data-test-range-type="max"
-                  fullWidth
-                  name={filter.name + "_max"}
-                  InputProps={{
-                    classes: {
-                      input: classes.input
-                    },
-                    endAdornment:
-                      filter.type === FieldType.price && currencySymbol,
-                    type: filter.type === FieldType.date ? "date" : "number"
-                  }}
-                  value={filter.value[1]}
-                  onChange={event =>
-                    onFilterPropertyChange({
-                      payload: {
-                        name: filter.name,
-                        update: {
-                          value: [filter.value[0], event.target.value]
-                        }
-                      },
-                      type: "set-property"
-                    })
-                  }
-                />
-              </>
+              <FilterRangeField
+                currencySymbol={currencySymbol}
+                filter={filter}
+                onFilterPropertyChange={onFilterPropertyChange}
+              />
             ) : (
-              <TextField
-                data-test={filterTestingContext}
-                data-test-id={filter.name}
-                fullWidth
-                name={filter.name}
-                InputProps={{
-                  classes: {
-                    input: classes.input
-                  },
-                  endAdornment:
-                    filter.type === FieldType.price && currencySymbol,
-                  type:
-                    filter.type === FieldType.date
-                      ? "date"
-                      : [FieldType.number, FieldType.price].includes(
-                          filter.type
-                        )
-                      ? "number"
-                      : "text"
-                }}
-                value={filter.value[0]}
-                onChange={event =>
-                  onFilterPropertyChange({
-                    payload: {
-                      name: filter.name,
-                      update: {
-                        value: [event.target.value, filter.value[1]]
-                      }
-                    },
-                    type: "set-property"
-                  })
-                }
+              <FilterTextField
+                currencySymbol={currencySymbol}
+                filter={filter}
+                onFilterPropertyChange={onFilterPropertyChange}
               />
             )}
           </div>
